Fix FireTOS booting
For some reason which I coudn't figure it out FireTOS
fails to boot randomly when we turn on the MMU before
the IDE ports are initialized (FireTOS doesn't detect
the IDE devices). This is fixed when we turn on the MMU
after the IDE initialization. CodeWarrior BaS does it
this way too.